+++ lib/src/widgets/link_view_horizontal.dart
@@ -159,10 +159,8 @@ class LinkViewHorizontal extends StatelessWidget {
                   description,
                   textAlign: TextAlign.left,
                   style: _bodyTS,
-                  overflow: bodyTextOverflow == null
-                      ? TextOverflow.ellipsis
-                      : bodyTextOverflow,
-                  maxLines: bodyMaxLines == null ? _maxLines : bodyMaxLines,
+                  overflow: bodyTextOverflow ?? TextOverflow.ellipsis,
+                  maxLines: bodyMaxLines ?? _maxLines,
                 ),
               ),
             ),